#1272 test: probe を tier1-2 の回帰テストへ昇格する
probe（scope=probe / iOS）が run 31730543074 で緑になり、#1272 の 2 層の修正
（5b46f082: クエリを起動リダイレクトで落とさない / 9cd3854d: jumpToTab を
auth 解決に追従させる）が確認できたため、tests/probe/README.md の手順 4 に従い
tests/profile/profile-tab-deep-link.test.ts へ昇格する（describeProbe → describeAuthenticated）。

観測点は見た目ではなくアプリ側プローブ（routeParamsProbe）のまま残す。
「クエリが届いたか」と「タブ切り替えが発火したか」は別の層の事実で、
見た目だけを見ると退行時にどの層が壊れたのかがまた分からなくなる
（観測なしで 3 仮説を外し続けた反省）。失敗メッセージは層まで名指しする。

あわせて「?tab= は効かない・原因未特定」と書いていた 2 箇所のコメントを
「#1272 で修正済み・回帰は profile-tab-deep-link.test.ts が担保」へ更新した。
saved-topic-location-search の入口はテーマが地点検索なので実 UI のままにする。
